Handle floats and fix string allocations in obituary text handler.

+++ p_view.c
@@ -1090,7 +1090,7 @@ void ClientEndServerFrame (edict_t *ent)
 	// calculate speed and cycle to be used for
 	// all cyclic walking effects
 	//
-	xyspeed = sqrt(ent->velocity[0]*ent->velocity[0] + ent->velocity[1]*ent->velocity[1]);
+	xyspeed = sqrtf(ent->velocity[0]*ent->velocity[0] + ent->velocity[1]*ent->velocity[1]);
 
 	// Expert: If the hook is used, no use/want for bobbing.
 	// Expert: Don't bob in a pause either
